PLAT-863 | Enable tag deletion

+++ app/Http/Controllers/Api/PrivateApi/TagsApiController.php
@@ -3,7 +3,6 @@
 use App\Http\Controllers\Api\ApiController;
 use App\Http\Requests\Course\UpdateTag;
 use App\Models\Tag;
-use App\Models\Taggable;
 use Illuminate\Http\Request;
 
 class TagsApiController extends ApiController {
@@ -61,9 +60,6 @@ class TagsApiController extends ApiController {
 	public function delete($id) {
 		$tag = Tag::find($id);
 
-		// TODO
-		return $this->respondNotImplemented();
-
 		if (!$tag) {
 			return $this->respondNotFound();
 		}
@@ -85,6 +81,14 @@ class TagsApiController extends ApiController {
 				'message' => __('tags.errors.in-course-structure')
 			]);
 		}
+
+		// Last defense
+		if (!$tag->isDeleteAllowed()) {
+			return $this->respondUnprocessableEntity([
+				'message' => __('tags.errors.delete-disallowed')
+			]);
+		}
+
 
 		$tag->delete();
